Move cache out of ginkeycloak to own package for re-use
